Title: 배열의 평균, Time: 143ms, Memory: 10MB, Status: Passed - CodeTree

+++ 240913/배열의 평균/ave-of-array.java
@@ -5,8 +5,8 @@ public class Main {
         // 여기작성해주세요.
         int[][] x = new int[2][4];
         Scanner sc = new Scanner(System.in);
-        int[] ii = new int[2];
-        int[] jj = new int[4];
+        double[] ii = new double[2];
+        double[] jj = new double[4];
         int tsum = 0;
         int s = 0;
